feat(purge all): Include option to purge all the cached files
Origin: https://github.com/FRiCKLE/ngx_cache_purge/pull/35/commits/63a1014709310e9fc1ef64afb6025076abe03b36

This option can be slow if a lot of content is cached, or if the
storage used for the cache is slow. But you really should be using
RAM as your cache storage.

Signed-off-by: Francisco Miguel Biete <fbiete@gmail.com>
<fmbiete@gmail.com>

Gbp-Pq: Name 0017-feat-purge-all-Include-option-to-purge-all-the-cache.patch