Proxy: add "proxy_ssl_alpn" directive.
Origin: http://mailman.nginx.org/pipermail/nginx-devel/2017-June/010217.html

ALPN is used here only to indicate which version of the HTTP protocol
is going to be used and we doesn't verify that upstream agreed to it.

Please note that upstream is allowed to reject SSL connection with a
fatal "no_application_protocol" alert if it doesn't support it.

Signed-off-by: Piotr Sikora <piotrsikora@google.com>

Gbp-Pq: Name 0009-Proxy-add-proxy_ssl_alpn-directive.patch